pyproject.toml
[tool.poetry]
name = "jumpbookstore"
version = "0.1.0"
description = "downloader for jumpbookstore"
authors = ["Hiroki Ishioka <ishioka0222@gmail.com>"]

[tool.poetry.scripts]
jumpbookstore = "jumpbookstore.cli:main"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
requests = "^2.22"
tqdm = "^4.32"
pycryptodome = "^3.8"
pillow = "^6.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
mypy = "^0.720.0"
flake8 = "^3.7"
autopep8 = "^1.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
